+++ src/components/form/MultifieldGroup/multifieldGroup.js
@@ -72,6 +72,7 @@ function initGroup(group) {
       errorField.classList.remove('input-field--error');
       errorField.classList.remove('dropdown--error');
       errorField.classList.remove('textarea--error');
+      errorField.classList.remove('radio--error');
     }
 
     for (const i in dataItems) {
@@ -84,12 +85,13 @@ function initGroup(group) {
           for (const fieldName of line.errors) {
             const elements = item.querySelectorAll(`[data-field-name="${fieldName}"]`);
             for (const element of elements) {
-              if (element.tagName === 'INPUT' && element.closest('.input-field')) {
+              if (element.tagName === 'INPUT' && element.type === 'radio') {
+                element.closest('.radio').classList.add('radio--error');
+              } else if (element.tagName === 'INPUT' && element.closest('.input-field')) {
                 element.closest('.input-field').classList.add('input-field--error');
               } else if (element.tagName === 'SELECT') {
                 element.closest('.dropdown').classList.add('dropdown--error');
-              }
-              if (element.tagName === 'TEXTAREA') {
+              } else if (element.tagName === 'TEXTAREA') {
                 element.closest('.textarea').classList.add('textarea--error');
               }
             }
@@ -144,7 +146,12 @@ function addGroup(group, values = { values: {}, errors: [], placeholders: {} }) 
           addField(multivalueField, value);
         }
       } else {
-        if (element.tagName === 'INPUT') {
+        if (element.tagName === 'INPUT' && element.type === 'radio') {
+          const defaultRadioElem = template.querySelector(`[name^="${key}"][value="${values.values[key]}"]`);
+          if (defaultRadioElem) {
+            defaultRadioElem.checked = true;
+          }
+        } else if (element.tagName === 'INPUT') {
           element.setAttribute('value', values.values[key]);
         } else if (element.tagName === 'TEXTAREA') {
           element.innerHTML = values.values[key];
@@ -176,8 +183,15 @@ function addGroup(group, values = { values: {}, errors: [], placeholders: {} }) 
     // Set mapping between labels and fields for single fields.
     if (element.tagName === 'INPUT') {
       const multiWrapper = element.closest('.multivalue-field');
-      if (!multiWrapper) {
+
+      if (!multiWrapper && element.type !== 'radio') {
         const singleWrapper = element.closest('.input-field');
+        if (singleWrapper) {
+          const label = singleWrapper.querySelector('label');
+          setLabelMappingForInput(label);
+        }
+      } else if (!multiWrapper && element.type === 'radio') {
+        const singleWrapper = element.closest('.radio');
         if (singleWrapper) {
           const label = singleWrapper.querySelector('label');
           setLabelMappingForInput(label);